+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+# 1.0.69 (2025-10-19)
+
+-   Improve email validation in login form with regex
+-   Add server headers for security
+-   Add global error handling middleware in Express
+-   Add uncaught exception handler in server.ts
+
 # 1.0.68 (2025-10-12)
 
 -   Cleanup code